+++ rtvi-client-js/src/index.ts
@@ -4,5 +4,6 @@ export * from "./errors";
 export * from "./events";
 export * from "./helpers";
 export * from "./helpers/llm";
+export * from "./logger";
 export * from "./messages";
 export * from "./transport";