+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to **RTVI Client Web** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.2.2] - 2024-11-12
+
+### Added
+
+- `setLogLevel` method added to `RTVIClient` to allow developers to set the log level.
+- `logger` instance added to `RTVIClient` to allow developers to set the log level.
+
 ## [0.2.1] - 2024-10-28
 
 ### Removed